% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcapower.R
\name{qp_quant_plot}
\alias{qp_quant_plot}
\title{Plot of 5%-quantiles for assessing the dispersion of the permutated
distributions}
\usage{
qp_quant_plot(power_est, title = FALSE)
}
\arguments{
\item{power_est}{Dataframe containing simulation results (see
\code{\link{qcapower}})}

\item{title}{Option for adding title to plot (default \code{FALSE})}
}
\description{
Depending on the number of cases, the permuted distributions of consistency
values can differ widely in terms of their location on the spectrum and their
shape.
}
\details{
Creates a sina plot with \code{ggforce}
}
\examples{
sim_data <- qp_sina_data
qp_quant_plot(sim_data)
}
