% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qp_sina_data}
\alias{qp_sina_data}
\title{Data simulated power estimates for plotting of 5\%-quantiles}
\format{
A dataframe with 1000 rows and 6 variables:
}
\usage{
qp_sina_data
}
\description{
A dataset containing the estimated 5\%-quantiles from a power simulation with
1000 simulations each with 10000 permutations. The value for the alternative
hypothesis was set to 1.
}
\details{
\describe{
\item{power}{power estimate over 1000 simulations}
\item{powercum}{running power estimate for ith simulation}
\item{null_hypo}{null hypothesis (H0), set to 0.8 (irrelevant here)}
\item{alt_hypo}{alternative hypothesis (H1), set to 1}
\item{cases}{number of cases, set to 10}
\item{quant}{estimated 5\%-quantiles per simulations}
}
}
\keyword{datasets}
