% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qp_sim_power}
\alias{qp_sim_power}
\title{Data simulated power estimates}
\format{A dataframe with simulation parameters and calculated power}
\usage{
qp_sim_power
}
\description{
A dataset containing power simulations for different number of cases and
different values for null- and alternative hypothesis
}
\details{
\describe{
\item{cases}{number of cases}
\item{null_hypo}{null hypothesis (H0)}
\item{alt_hypo}{alternative hypothesis}
\item{sims}{number of simulations}
\item{perms}{number of permutations}
\item{perms}{calculate power}
}
}
\keyword{datasets}
